Created baseline 1 month prediction model.
Using our 1 month DataFrame, we created a baseline model using the sameparameters as our 1 week baseline model. With our 1 week, we only trained it on data before feb 10 2024 and tested it on unseen data after feb 17 2024. With our 1 month model, we changed it to train on only data before Jan 10 2024 and test it on unseen data after Feb 10 2024. Then we printed metrics and features.