Removed 1 year prediction moodel and added visual graphs.
My 1 year model did not have enough data to predict 1 year ahead, so the metrics were off and it had to be removed until I gain access to more data in a separate project. On top of this, I hard coded line graphs to show the change in metric values as the timeframe prediction increases. Visualzed metrics: Root Mean Squared, R^2, MAPE, and Median Absolute Error.