Added function to print out predicted price and actual price of testing data.
I wrote a function previously to have the user input a stock ticker and a date, but some dates get removed during the model training. To get by this, i first created a function to print out available dates which have an actual price and predicted price in the row.