Created two functions for outputting date and predicted price.
Created two funtions. The first function allows the user to enter a stock ticker from Symbol column and the function prints out the first available date for prediction in the test data and the last available date for prediction in the test date. The second fuction allows user to input Symbol and date and function outputs actual price and predicted price.