Created trading simulation to backtest final model and performed some statistical calculations to measure the actual predictive power of the model.
Once our final model was chosen, I created a new dataframe containing the predicted values for the last two years of testing data, along with the actual values, and the value of the stock 1 week prior that the prediction was based on. Using these 3 columns, I was able to test the performance of the model  in an interpretable way, as well as create a trading simulation to test what it would be like to trade stocks using this model, measuring how much money was gained or lost over the two years of testing data.